# docs/api-spec/erp-services-api-openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: support@example.com
    name: ERP API Team
  description: API specification for ERP resource planning and management.
  title: ERP Services API
  version: 1.0.0
servers:
- description: Production
  url: https://api.example.com/v1
- description: Sandbox
  url: https://sandbox.example.com/v1
security:
- bearerAuth: []
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      parameters:
      - explode: true
        in: query
        name: accountId
        required: false
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: startDate
        required: false
        schema:
          format: date
          type: string
        style: form
      - explode: true
        in: query
        name: endDate
        required: false
        schema:
          format: date
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Transaction"
                type: array
          description: Transaction list
      summary: List transactions
      tags:
      - Transactions
      x-accepts: application/json
      x-tags:
      - tag: Transactions
    post:
      operationId: createTransaction
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
          description: Transaction created
      summary: Create transaction
      tags:
      - Transactions
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Transactions
...
components:
  schemas:
...
    Transaction:
      example:
        transactionType: DEBIT
        accountId: accountId
        amount: 0.8008281904610115
        description: description
        currency: currency
        transactionDate: 2000-01-23T04:56:07.000+00:00
        transactionId: transactionId
      properties:
        transactionId:
          type: string
        accountId:
          type: string
        amount:
          format: double
          type: number
        currency:
          type: string
        transactionType:
          enum:
          - DEBIT
          - CREDIT
          type: string
        description:
          type: string
        transactionDate:
          format: date-time
          type: string
      type: object
    CreateTransactionRequest:
      example:
        transactionType: transactionType
        accountId: accountId
        amount: 0.8008281904610115
        description: description
      properties:
        accountId:
          type: string
        amount:
          format: double
          type: number
        transactionType:
          type: string
        description:
          type: string
      required:
      - accountId
      - amount
      - transactionType
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http